VIH-9395 - moved logic to domain service
