--- BookingsApi/BookingsApi/Services/IHearingBusiness.cs
@@ -1,12 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using BookingsApi.Domain;
-
-namespace BookingsApi.Services
-{ 
-    public interface IHearingBusiness
-    {
-        Task<List<VideoHearing>> GetUnallocatedHearings();
-    }
-}
